crypto_extras.py
#!/usr/bin/env python3
"""
Created by Tanmay Patil
Copyright © 2025 Tanmay Patil. All rights reserved.

This module provides additional cryptographic features such as digital signatures,
advanced hashing, MAC, password hashing/derivation, and public-key operations.
"""

import hashlib
from Crypto.Hash import Poly1305
from Crypto.Cipher import AES
from nacl.signing import SigningKey, VerifyKey
from nacl.exceptions import BadSignatureError
from argon2 import PasswordHasher
from cryptography.hazmat.primitives.kdf.scrypt import Scrypt
from cryptography.hazmat.backends import default_backend
from cryptography.hazmat.primitives import serialization
from cryptography.hazmat.primitives.asymmetric import x25519
import os

def ed25519_sign(message: bytes, private_key_hex: str) -> bytes:
    private_key = bytes.fromhex(private_key_hex)
    signing_key = SigningKey(private_key)
    signed = signing_key.sign(message)
    return signed.signature

def ed25519_verify(message: bytes, signature: bytes, public_key_hex: str) -> bool:
    public_key = bytes.fromhex(public_key_hex)
    verify_key = VerifyKey(public_key)
    try:
        verify_key.verify(message, signature)
        return True
    except BadSignatureError:
        return False

def blake2b_hash(message: bytes) -> bytes:
    h = hashlib.blake2b()
    h.update(message)
    return h.digest()

def shake256_hash(message: bytes, digest_size: int = 64) -> bytes:
    h = hashlib.shake_256()
    h.update(message)
    return h.digest(digest_size)

def hongjun_hash(message: bytes) -> bytes:
    # Placeholder implementation (in a real system, integrate a proven Hongjun library)
    # For demonstration, we simply return a SHA-256 hash prefixed with b"HJ"
    return b"HJ" + hashlib.sha256(message).digest()

def ripemd320_hash(message: bytes) -> bytes:
    # Since RIPEMD-320 is not available in standard libraries, we simulate it by using RIPEMD-160 twice.
    # This is only for demonstration purposes.
    from Crypto.Hash import RIPEMD
    h1 = RIPEMD.new(message, digest_bits=160).digest()
    h2 = RIPEMD.new(message[::-1], digest_bits=160).digest()
    return (h1 + h2)[:40]  # 320 bits = 40 bytes

def poly1305_mac(message: bytes, key: bytes) -> bytes:
    mac = Poly1305.new(key=key, cipher=AES)
    mac.update(message)
    return mac.digest()

def argon2_hash(password: str, salt: bytes, type: str = "argon2id") -> str:
    ph = PasswordHasher()  # defaults to argon2id
    return ph.hash(password + salt.hex())

def scrypt_derive(password: str, salt: bytes) -> bytes:
    kdf = Scrypt(
        salt=salt,
        length=32,
        n=2**14,
        r=8,
        p=1,
        backend=default_backend()
    )
    return kdf.derive(password.encode())

def generate_x25519_key_pair() -> tuple[bytes, bytes]:
    private_key = x25519.X25519PrivateKey.generate()
    public_key = private_key.public_key()
    priv_bytes = private_key.private_bytes(
        encoding=serialization.Encoding.Raw,
        format=serialization.PrivateFormat.Raw,
        encryption_algorithm=serialization.NoEncryption()
    )
    pub_bytes = public_key.public_bytes(
        encoding=serialization.Encoding.Raw,
        format=serialization.PublicFormat.Raw
    )
    return priv_bytes, pub_bytes

def x25519_derive_shared(private_key_bytes: bytes, peer_public_key_bytes: bytes) -> bytes:
    private_key = x25519.X25519PrivateKey.from_private_bytes(private_key_bytes)
    peer_public_key = x25519.X25519PublicKey.from_public_bytes(peer_public_key_bytes)
    shared_key = private_key.exchange(peer_public_key)
    return shared_key

def threshold_sign(message: bytes, private_keys_hex: list) -> bytes:
    # Placeholder multisignature: simply concatenate individual Ed25519 signatures.
    sigs = []
    for key_hex in private_keys_hex:
        sig = ed25519_sign(message, key_hex)
        sigs.append(sig)
    # In a real threshold scheme, signatures would be combined securely.
    return b"".join(sigs)

def threshold_verify(message: bytes, combined_signature: bytes, public_keys_hex: list) -> bool:
    # Placeholder: split the combined signature evenly and verify each one.
    sig_len = 64  # Ed25519 signature length
    if len(combined_signature) % sig_len != 0:
        return False
    num = len(combined_signature) // sig_len
    if num != len(public_keys_hex):
        return False
    for i, pub_hex in enumerate(public_keys_hex):
        sig = combined_signature[i*sig_len:(i+1)*sig_len]
        if not ed25519_verify(message, sig, pub_hex):
            return False
    return True
