hyperparameter_tuner.py
#!/usr/bin/env python3
"""
Created by Tanmay Patil
Copyright © 2025 Tanmay Patil. All rights reserved.

This module uses Bayesian Optimization to tune encryption hyperparameters.
The simulated function is designed to mimic encryption performance while balancing speed and security.
"""

from bayes_opt import BayesianOptimization
import time
import random


def encryption_performance(iteration_count, block_size):
    """
    Simulate encryption time as a function of iteration_count and block_size.
    Lower simulated time yields a higher reward.
    """
    # Use a simplified simulation that is fast to compute
    simulated_time = (iteration_count / 100000) * (block_size / 32) * random.uniform(0.8, 1.2)
    return -simulated_time  # Lower time => higher reward


def run_tuning():
    pbounds = {
        'iteration_count': (50000, 150000),
        'block_size': (16, 128)
    }
    optimizer = BayesianOptimization(
        f=encryption_performance,
        pbounds=pbounds,
        random_state=1,
        verbose=2
    )
    print("Starting Bayesian Optimization for hyperparameter tuning...")
    optimizer.maximize(init_points=3, n_iter=5)  # Fewer iterations for low-end hardware
    print("Tuning complete. Best parameters:")
    print(optimizer.max)


if __name__ == "__main__":
    run_tuning()
