requirements.txt
pycryptodome
cryptography
argparse
tqdm
rich
prometheus_client
alive-progress
PyNaCl
argon2-cffi
bayesian-optimization
scikit-learn
xgboost
lightgbm
joblib
transformers
torch
numpy
pyreadline3
readchar
prompt_toolkit
PyYAML